+++ lib/model/cross_file.dart
@@ -1,5 +1,9 @@
+import 'dart:typed_data';
+
+import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:localsend_app/model/file_type.dart';
+import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 
 part 'cross_file.freezed.dart';
 
@@ -10,7 +14,19 @@ class CrossFile with _$CrossFile {
     required String name,
     required FileType fileType,
     required int size,
+    required AssetEntity? asset, // for thumbnails
     required String? path,
-    required List<int>? bytes,
+    required Future<Uint8List> Function()? bytes,
   }) = _CrossFile;
+}
+
+extension AssetEntityExt on AssetEntity {
+  Widget get thumbnailWidget {
+    return AssetEntityImage(
+      this,
+      isOriginal: false,
+      thumbnailSize: const ThumbnailSize.square(32),
+      thumbnailFormat: ThumbnailFormat.jpeg,
+    );
+  }
 }