+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- feat(android): add media picker
+- feat(iOS): merge image and video to common media picker
+
 ## 1.0.0 (2022-12-29)
 
 - Initial Release